// AcoPlaceFrontend/src/app/model/announcement-details.ts
export class AnnouncementDetails {
    id: number;
    bath: boolean; 
    shower: boolean; 
    microwave: boolean; 
    oven: boolean; 
    petsAllowed: boolean; 
    elevator: boolean; 
    nearPark: boolean; 
    fenced: boolean; 
    nearTram: boolean; 
    nearBus: boolean;
    wifi: boolean; 
    ethernetOutlets: boolean; 
    internet: boolean;
    internetSpeed: number;
    tv: boolean; 
    dishwasher: boolean;
    clothesDryer: boolean;
    nearShoppingMall: boolean;
    nearBakery: boolean;
    nearFoodMarket: boolean;
    nearSupermarket: boolean; 

}